+++ lib/src/integrity/integrity_checker.dart
@@ -289,6 +289,8 @@ class IntegrityChecker {
         return value is bool;
       case DataType.datetime:
         return value is String && DateTime.tryParse(value) != null;
+      case DataType.array:
+        return value is List;
     }
   }
 }